Improve random strategy & add total assets value calculation

+++ binance_bot/strategy/strategy_action.py
@@ -1,11 +1,9 @@
 from typing import Literal
 
-import numpy as np
-
 
 class StrategyAction:
 
-    def __init__(self, side: Literal["SELL", "BUY"], pair: str, quantity: np.double):
+    def __init__(self, side: Literal["SELL", "BUY"], pair: str, quantity: float):
         self.side = side
         self.pair = pair
         self.quantity = quantity
